Flatten Python structure, more SCSS cleanup
