--- content/README.md
@@ -1,5 +1,0 @@
----
-home: true
-title: "Florimond Manca"
-description: "A blog about my ongoing journey through software engineering."
----
