+++ tests/test_app.py
@@ -103,20 +103,3 @@ async def test_meta(client: httpx.AsyncClient) -> None:
     assert find_meta("property", "og:title") is not None
     assert find_meta("property", "og:description") is not None
     assert find_meta("property", "og:image") is not None
-
-
-@pytest.mark.parametrize(
-    "resource",
-    [
-        "/static/fonts/eb-garamond/regular.ttf",
-        "/static/fonts/eb-garamond/medium.ttf",
-        "/static/fonts/eb-garamond/bold.ttf",
-        "/static/fonts/eb-garamond/italic.ttf",
-    ],
-)
-async def test_caching(client: httpx.AsyncClient, resource: str) -> None:
-    url = f"http://florimond.dev{resource}"
-    resp = await client.get(url, allow_redirects=False)
-    assert resp.status_code == 200, resp.url
-    assert "cache-control" in resp.headers
-    assert "max-age=3600" in resp.headers["cache-control"]
