--- www/blog/__init__.py
@@ -1,4 +1,0 @@
-from .event_handlers import on_shutdown, on_startup
-from .routes import routes
-
-__all__ = ["on_startup", "on_shutdown", "routes"]
