+++ www/templates/base.jinja
@@ -59,7 +59,7 @@
     <script src="{{ url_for('static', path='/js/ga.js') }}"></script>
 
     {% if settings.DEBUG %}
-      {{ hotreload.script(url_for('blog:hot-reload')) | safe }}
+      {{ hotreload.script(url_for('hot-reload')) | safe }}
     {% endif %}
 
     {% block tail %}{% endblock %}
