+++ www/templates/components/jumbotron.jinja
@@ -20,11 +20,6 @@
 
   <ul class="c-jumbotron-links u-list-reset">
     <li>
-      <a href="{{ url_for('blog:home') }}">
-        ✍️&nbsp;&nbsp;Writings
-      </a>
-    </li>
-    <li>
       <a href="https://github.com/florimondmanca" target="_blank" rel="noopener">
         💻&nbsp;&nbsp;Code
       </a>
