+++ www/templates/components/pages/tag_list.jinja
@@ -2,7 +2,7 @@
 <ul class="u-list-reset {{ class }} {{ 'u-flex-inline' if inline else '' }}">
   {% for tag in tags %}
   <li class="u-mh-1">
-    <a href="{{ url_for('blog', path='/tag/' + tag) }}">
+    <a href="{{ url_for('page', permalink='tag/' + tag) }}">
       #{{ tag }}
     </a>
   </li>
